use size_t, fix warnings under -Wstrict_prototypes and more

+++ trunk/src/modp_b2.h
@@ -37,7 +37,7 @@ BEGIN_C
  * \param[in] len of the input string
  * \return strlen of dest
  */
-int modp_b2_encode(char* dest, const char* str, int len);
+size_t modp_b2_encode(char* dest, const char* str, size_t len);
 
 /**
  * Decode a hex-encoded string.
@@ -46,9 +46,9 @@ int modp_b2_encode(char* dest, const char* str, int len);
  *   input must be a mutliple of 2, and be different than the source buffer.
  * \param[in] src the hex encoded source
  * \param[in] len the length of the source
- * \return the length of the the output, or -1 if an error
+ * \return the length of the the output, or 0 if an error (input size not a multiple of 8)
  */
-int modp_b2_decode(char* dest, const char* src, int len);
+size_t modp_b2_decode(char* dest, const char* src, size_t len);
 
 /**
  * Encode length.
